Created "Add Item" function: Creates JSON structure for actor reference.
Added a new function that can be called anywhere. It adds an items information to a JSON file just given its object reference. Next step is to allow it to add items as children and add items with children. Also a function for moving items between JSON files. (World / Player)